// types/eve.ts
export interface eveCorp {
    eve_online_id: string;
    eve_corp_id: string;
    char_name: string;
    picture: string;
    corp_logo: string;
    corp_name: string;
    corp_member_count: string;
    corp_ticker: string;
    corp_top_list: string;
}

export interface eveMember {
    eve_online_id: string;
    eve_corp_id: string;
    char_name: string;
    picture: string;
    corp_logo: string;
}